Fix Vue design regressions in navigation
