+++ src/views/AccountSettings.vue
@@ -1,19 +1,6 @@
 <template>
 	<AppContent app-name="mail">
-		<template slot="navigation">
-			<AppNavigationNew :text="t('mail', 'New message')"
-							  buttonId="mail_new_message"
-							  buttonClass="icon-add"
-							  @click="onNewMessage"/>
-			<ul id="accounts-list">
-				<AppNavigationItem v-for="item in menu"
-								   :key="item.key"
-								   :item="item"/>
-			</ul>
-			<AppNavigationSettings :title="t('mail', 'Settings')">
-				<AppSettingsMenu />
-			</AppNavigationSettings>
-		</template>
+		<Navigation slot="navigation" />
 		<template slot="content">
 			TODO: implement account settings
 		</template>
@@ -21,24 +8,16 @@
 </template>
 
 <script>
-	import {
-		AppContent,
-		AppNavigationItem,
-		AppNavigationNew,
-		AppNavigationSettings
-	} from 'nextcloud-vue'
-	import AppSettingsMenu from '../components/AppSettingsMenu'
+	import {AppContent} from 'nextcloud-vue'
 
-	import SidebarItems from '../mixins/SidebarItems'
+	import AppSettingsMenu from '../components/AppSettingsMenu'
+	import Navigation from '../components/Navigation'
 
 	export default {
 		name: 'AccountSettings',
-		extends: SidebarItems,
 		components: {
 			AppContent,
-			AppNavigationItem,
-			AppNavigationNew,
-			AppNavigationSettings,
+			Navigation,
 			AppSettingsMenu,
 		},
 		computed: {
